creating a new lightweight API controller
